feat: enhance resource inheritance functionality
- Introduced environment variable support for inheriting Kubernetes resources and nodeSelector in the GetTranslations function.
- Replaced direct calls to model functions with environment variable checks for better configurability.
- Added a new helper function, getInheritedResourcesFromContainer, to streamline resource inheritance logic.
- Implemented comprehensive unit tests for the new inheritance behavior, ensuring correct functionality across various scenarios.
